+++ src/main/java/spot/spot/domain/job/command/mapper/WorkerCommandMapper.java
@@ -1,4 +1,4 @@
-package spot.spot.domain.job.mapper;
+package spot.spot.domain.job.command.mapper;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -6,11 +6,11 @@ import java.util.stream.Collectors;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.ReportingPolicy;
-import spot.spot.domain.job.dto.Location;
-import spot.spot.domain.job.dto.request.RegisterWorkerRequest;
-import spot.spot.domain.job.dto.response.NearByJobResponse;
-import spot.spot.domain.job.entity.Job;
-import spot.spot.domain.job.util.JobUtil;
+import spot.spot.domain.job.command.dto.Location;
+import spot.spot.domain.job.command.dto.request.RegisterWorkerRequest;
+import spot.spot.domain.job.query.dto.response.NearByJobResponse;
+import spot.spot.domain.job.command.entity.Job;
+import spot.spot.domain.job.command.util.JobUtil;
 import spot.spot.domain.member.entity.Ability;
 import spot.spot.domain.member.entity.AbilityType;
 import spot.spot.domain.member.entity.Member;
@@ -21,27 +21,14 @@ import spot.spot.domain.member.repository.AbilityRepository;
 
 @Mapper(componentModel = "spring",
         unmappedTargetPolicy = ReportingPolicy.IGNORE, uses = {AbilityRepository.class, JobUtil.class})
-public interface WorkerMapper {
-
-    /*
-    *   1) 구직자 등록 (request -> entity)
-    *   2) 근처 일 하나 변환 (entity -> response)
-    *   3) Worker가 입력한 자신의 능력들과 Worker 본인을 교차테이블로 연관관계 짓기 (default)
-    *   4) 2번을 List로 변환
-    * */
+public interface WorkerCommandMapper {
 
     // 1) Registing request to Entity
     @Mapping(target = "member", source = "member")
     @Mapping(target = "introduction", source = "request.content")
     @Mapping(target = "workerAbilities", ignore = true) // WorkerAbility 매핑은 별도 처리
     Worker dtoToWorker(RegisterWorkerRequest request, Member member);
-    // 2) 근처 일 하나 변환 (entity to response)
-    @Mapping(target = "dist", ignore = true)
-    NearByJobResponse toNearByJobResponse(Job job);
-
-
-
-    // 3) Worker가 입력한 자신의 능력들과 Worker 본인을 교차테이블로 연관관계 짓기 (default)
+    // 2) 구직자와 강점의 교차테이블 생성
     default List<WorkerAbility> mapWorkerAbilities(List<AbilityType> strong, Worker worker, AbilityRepository abilityRepository) {
         if(strong == null || strong.isEmpty()) return new ArrayList<>();
         return strong.stream()
@@ -55,15 +42,5 @@ public interface WorkerMapper {
                     .build();
             })
             .collect(Collectors.toList());
-    }
-    // 4) 2번을 리스트로 변환
-    default List<NearByJobResponse> toNearByJobResponseList(List<Job> jobs, Location location) {
-        return jobs.stream()
-            .map(job -> {
-                NearByJobResponse response = toNearByJobResponse(job);
-                double distance = JobUtil.calculateHaversineDistance(location.lat(), location.lng(), job.getLat(), job.getLng());
-                return response.toBuilder().dist(distance).build(); //기존 객체를 복사 + dist만 변경
-            })
-            .toList();
     }
 }
