+++ src/main/java/spot/spot/domain/job/command/service/WorkerCommandService.java
@@ -1,35 +1,23 @@
-package spot.spot.domain.job.service;
+package spot.spot.domain.job.command.service;
 
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
-import spot.spot.domain.job.dto.response.JobCertifiationResponse;
-import spot.spot.domain.job.mapper.WorkerMapper;
-import spot.spot.domain.job.util.JobUtil;
-import spot.spot.domain.job.dto.request.ChangeStatusWorkerRequest;
-import spot.spot.domain.job.dto.request.RegisterWorkerRequest;
-import spot.spot.domain.job.dto.request.YesOrNoClientsRequest;
-import spot.spot.domain.job.dto.response.JobDetailResponse;
-import spot.spot.domain.job.dto.response.JobSituationResponse;
-import spot.spot.domain.job.dto.response.NearByJobResponse;
-import spot.spot.domain.job.entity.Certification;
-import spot.spot.domain.job.entity.Job;
-import spot.spot.domain.job.entity.Matching;
-import spot.spot.domain.job.entity.MatchingStatus;
-import spot.spot.domain.job.repository.dsl.ChangeJobStatusDsl;
-import spot.spot.domain.job.repository.dsl.MatchingDsl;
-import spot.spot.domain.job.repository.dsl.SearchingListDsl;
-import spot.spot.domain.job.repository.jpa.CertificationRepository;
-import spot.spot.domain.job.repository.jpa.JobRepository;
-import spot.spot.domain.job.repository.jpa.MatchingRepository;
-import spot.spot.domain.job.service.searching.SearchingJobJPQLService;
-import spot.spot.domain.job.service.searching.SearchingJobNativeQueryService;
-import spot.spot.domain.job.service.searching.SearchingJobQueryDSLService;
+import spot.spot.domain.job.command.dto.response.JobCertifiationResponse;
+import spot.spot.domain.job.command.mapper.WorkerCommandMapper;
+import spot.spot.domain.job.command.util.JobUtil;
+import spot.spot.domain.job.command.dto.request.ChangeStatusWorkerRequest;
+import spot.spot.domain.job.command.dto.request.RegisterWorkerRequest;
+import spot.spot.domain.job.command.dto.request.YesOrNoClientsRequest;
+import spot.spot.domain.job.command.entity.Certification;
+import spot.spot.domain.job.command.entity.Job;
+import spot.spot.domain.job.command.entity.Matching;
+import spot.spot.domain.job.command.entity.MatchingStatus;
+import spot.spot.domain.job.command.repository.dsl.ChangeJobStatusCommandDsl;
+import spot.spot.domain.job.command.repository.jpa.CertificationRepository;
+import spot.spot.domain.job.query.repository.jpa.MatchingRepository;
 import spot.spot.domain.member.entity.Member;
 import spot.spot.domain.member.entity.Worker;
 import spot.spot.domain.member.repository.AbilityRepository;
@@ -48,61 +36,36 @@ import spot.spot.global.util.AwsS3ObjectStorage;
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class WorkerService {
-
-    // 거리 계산용 3가지
-    private final SearchingJobJPQLService searchingJobJPQLService;
-    private final SearchingJobNativeQueryService jobSearchNativeService;
-    private final SearchingJobQueryDSLService jobSearchQueryDSLService;
+public class WorkerCommandService {
     // Util
     private final UserAccessUtil userAccessUtil;
     private final FcmUtil fcmUtil;
-    private final WorkerMapper workerMapper;
+    private final WorkerCommandMapper workerCommandMapper;
     // Repo
     private final WorkerRepository workerRepository;
     private final AbilityRepository abilityRepository;
     private final WorkerAbilityRepository workerAbilityRepository;
-    private final JobRepository jobRepository;
     private final MatchingRepository matchingRepository;
     private final CertificationRepository certificationRepository;
-    private final ChangeJobStatusDsl changeJobStatusDsl;
+    private final ChangeJobStatusCommandDsl changeJobStatusCommandDsl;
     private final PayService payService;
     private final JobUtil jobUtil;
     private final AwsS3ObjectStorage awsS3ObjectStorage;
-    private final MatchingDsl matchingDsl;
-    private final SearchingListDsl searchingListDsl;
 
     @Transactional
     public void registeringWorker(RegisterWorkerRequest request) {
         Member member = userAccessUtil.getMember();
-        Worker worker = workerMapper.dtoToWorker(request, member);
+        Worker worker = workerCommandMapper.dtoToWorker(request, member);
         workerRepository.save(worker);
         workerAbilityRepository.saveAll(
-            workerMapper.mapWorkerAbilities(request.strong(), worker, abilityRepository));
+            workerCommandMapper.mapWorkerAbilities(request.strong(), worker, abilityRepository));
     }
     // 구직자 근처 일 리스트 반환
-    public Slice<NearByJobResponse> getNearByJobList(String impl, Double lat, Double lng, int zoom, Pageable pageable) {
-        Member member = userAccessUtil.getMember();
-        lat = lat == null? member.getLat() : lat;
-        lng = lng == null? member.getLng() : lng;
 
-        SearchingJobService service = switch (impl.toLowerCase()) {
-            case "jpql" -> searchingJobJPQLService;
-            case "native" -> jobSearchNativeService;
-            case "dsl" -> jobSearchQueryDSLService;
-            default -> throw new GlobalException(ErrorCode.INVALID_SEARCH_METHOD);
-        };
-        return service.findNearByJobs(lat, lng, zoom, pageable);
-    }
-    // 일 하나 상세 확인
-    public JobDetailResponse getOneJob (long jobId) {
-        Member me = userAccessUtil.getMember();
-        return matchingDsl.findOneJobDetail(jobId, me.getId()).orElseThrow(() -> new GlobalException(ErrorCode.MEMBER_NOT_FOUND));
-    }
     // 일 신청하기
     public void askingJob2Client(ChangeStatusWorkerRequest request) {
         Member worker = userAccessUtil.getMember();
-        Job job = changeJobStatusDsl.findJobWithValidation(worker.getId(), request.jobId());
+        Job job = changeJobStatusCommandDsl.findJobWithValidation(worker.getId(), request.jobId());
         Matching matching = Matching.builder().job(job).member(worker).status(MatchingStatus.ATTENDER).build();
         matchingRepository.save(matching);
         fcmUtil.singleFcmSend(worker.getId(), FcmDTO.builder().title("일 해결 신청 알림!").body(
@@ -112,10 +75,10 @@ public class WorkerService {
     @Transactional
     public void startJob (ChangeStatusWorkerRequest request) {
         Member worker = userAccessUtil.getMember();
-        Job job = changeJobStatusDsl.findJobWithValidation(worker.getId(), request.jobId(), MatchingStatus.YES);
+        Job job = changeJobStatusCommandDsl.findJobWithValidation(worker.getId(), request.jobId(), MatchingStatus.YES);
         PayHistory payHistory = payService.findByJob(job);
         payService.updatePayHistory(payHistory, PayStatus.PROCESS, worker.getNickname());
-        changeJobStatusDsl.updateMatchingStatus(worker.getId(), request.jobId(), MatchingStatus.START);
+        changeJobStatusCommandDsl.updateMatchingStatus(worker.getId(), request.jobId(), MatchingStatus.START);
         fcmUtil.singleFcmSend(worker.getId(), FcmDTO.builder().title("일 시작 알림!").body(
                 fcmUtil.getStartedJobMsg(worker.getNickname(), job.getTitle())).build());
     }
@@ -123,8 +86,8 @@ public class WorkerService {
     @Transactional
     public void yesOrNo2RequestOfClient(YesOrNoClientsRequest request) {
         Member worker = userAccessUtil.getMember();
-        Job job = changeJobStatusDsl.findJobWithValidation(worker.getId(), request.jobId(), MatchingStatus.REQUEST);
-        changeJobStatusDsl.updateMatchingStatus(worker.getId(), request.jobId(), request.isYes()? MatchingStatus.YES : MatchingStatus.NO);
+        Job job = changeJobStatusCommandDsl.findJobWithValidation(worker.getId(), request.jobId(), MatchingStatus.REQUEST);
+        changeJobStatusCommandDsl.updateMatchingStatus(worker.getId(), request.jobId(), request.isYes()? MatchingStatus.YES : MatchingStatus.NO);
         fcmUtil.singleFcmSend(worker.getId(), FcmDTO.builder().title("요청 승낙 알림!").body(
             fcmUtil.getStartedJobMsg(worker.getNickname(), job.getTitle())).build());
     }
@@ -134,7 +97,7 @@ public class WorkerService {
         Member worker = userAccessUtil.getMember();
         Matching matching = matchingRepository.findByMemberAndJob_Id(worker, request.jobId()).orElseThrow(() -> new GlobalException(ErrorCode.MATCHING_NOT_FOUND));
         jobUtil.withdrawalExistingScheduledTask(matching.getId());
-        changeJobStatusDsl.updateMatchingStatus(worker.getId(), request.jobId(), MatchingStatus.START);
+        changeJobStatusCommandDsl.updateMatchingStatus(worker.getId(), request.jobId(), MatchingStatus.START);
     }
 
     @Transactional
@@ -155,13 +118,8 @@ public class WorkerService {
         Matching matching = matchingRepository
             .findByMemberAndJob_Id(worker, request.jobId())
             .orElseThrow(() -> new GlobalException(ErrorCode.MATCHING_NOT_FOUND));
-        changeJobStatusDsl.findJobWithValidation(worker.getId(), request.jobId(), MatchingStatus.START, MatchingStatus.REJECT);
-        changeJobStatusDsl.updateMatchingStatus(worker.getId(), request.jobId(), MatchingStatus.FINISH);
-    }
-
-    public List<JobSituationResponse> getMyJobSituations() {
-        Member me = userAccessUtil.getMember();
-        return searchingListDsl.findJobSituationsByWorker(me.getId());
+        changeJobStatusCommandDsl.findJobWithValidation(worker.getId(), request.jobId(), MatchingStatus.START, MatchingStatus.REJECT);
+        changeJobStatusCommandDsl.updateMatchingStatus(worker.getId(), request.jobId(), MatchingStatus.FINISH);
     }
 
     @Transactional
