+++ src/main/java/spot/spot/domain/job/query/mapper/ClientQueryMapper.java
@@ -1,28 +1,21 @@
-package spot.spot.domain.job.mapper;
+package spot.spot.domain.job.query.mapper;
+
 
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
-
-import org.mapstruct.*;
-import spot.spot.domain.job.dto.request.RegisterJobRequest;
-import spot.spot.domain.job.dto.request.AttenderResponse;
-import spot.spot.domain.job.entity.Job;
+import org.mapstruct.IterableMapping;
+import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
+import org.mapstruct.Named;
+import org.mapstruct.ReportingPolicy;
+import spot.spot.domain.job.query.dto.request.AttenderResponse;
 import spot.spot.domain.member.entity.Worker;
 import spot.spot.domain.member.entity.WorkerAbility;
 
 @Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
-public interface ClientMapper {
-    /*
-    * 1) 일 등록 (reqeust -> entity)
-    * 2) 일 신청자 하나 변환 (entity -> response)
-    * 3) 2번의 리스트 형태
-    * 4) 3번 변환 시, 일에 필요한 능력을 교차테이블을 통해 String으로 변환하여 매핑하는 매소드
-    * */
+public interface ClientQueryMapper {
 
-    // 1) 일 등록  (reqeust -> entity)
-    @Mapping(source = "tid", target = "tid")
-    Job registerRequestToJob (String img, RegisterJobRequest request, String tid);
     // 2) 일 신청자 하나 변환 (entity -> response)
     @Mapping(source = "member.id", target = "id")
     @Mapping(source = "member.nickname", target = "name")
