+++ src/main/java/spot/spot/domain/job/query/util/searching/SearchingJobNativeQueryUtil.java
@@ -1,4 +1,4 @@
-package spot.spot.domain.job.service.searching;
+package spot.spot.domain.job.query.util.searching;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -6,21 +6,22 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.domain.SliceImpl;
 import org.springframework.stereotype.Service;
-import spot.spot.domain.job.mapper.WorkerMapper;
-import spot.spot.domain.job.dto.Location;
-import spot.spot.domain.job.dto.response.NearByJobResponse;
-import spot.spot.domain.job.entity.Job;
-import spot.spot.domain.job.repository.jpa.JobRepository;
-import spot.spot.domain.job.service.SearchingJobService;
-import spot.spot.domain.job.util.JobUtil;
+import spot.spot.domain.job.command.mapper.WorkerCommandMapper;
+import spot.spot.domain.job.command.dto.Location;
+import spot.spot.domain.job.query.dto.response.NearByJobResponse;
+import spot.spot.domain.job.command.entity.Job;
+import spot.spot.domain.job.query.mapper.WorkerQueryMapper;
+import spot.spot.domain.job.query.repository.jpa.JobRepository;
+import spot.spot.domain.job.query.util._docs.SearchingJobQueryUtil;
+import spot.spot.domain.job.command.util.JobUtil;
 
 @Service
 @RequiredArgsConstructor
-public class SearchingJobNativeQueryService implements SearchingJobService {
+public class SearchingJobNativeQueryUtil implements SearchingJobQueryUtil {
 
     private final JobRepository jobRepository;
     private final JobUtil jobUtil;
-    private final WorkerMapper workerMapper;
+    private final WorkerQueryMapper workerQueryMapper;
 
     @Override
     public Slice<NearByJobResponse> findNearByJobs(double lat, double lng, int zoom, Pageable pageable) {
@@ -34,7 +35,7 @@ public class SearchingJobNativeQueryService implements SearchingJobService {
             jobs = jobs.subList(0, pageable.getPageSize());
         }
 
-        List<NearByJobResponse> responseList = workerMapper
+        List<NearByJobResponse> responseList = workerQueryMapper
             .toNearByJobResponseList(jobs, Location.builder().lat(lat).lng(lng).build());
         return new SliceImpl<>(responseList, pageable, hasNext);
